requirements.txt
watchdog==6.0.0
pystray==0.19.5
pillow==11.2.1